Correct glyph and text bboxes in vertical writing mode (#110)
This is a less-refactored but more-tested version of #109

* fix: set default displacement vector correctly

* test: test PDF for complex vertical fonts

* fix!: correct totally bogus CID font terminology

Metrics are not entirely fixed yet but at least we are now referring
to things by their proper names and not using "width" for vertical
displacement for instance.

* feat: add .glyphs property to flatten harder

* fix: finally get vertical writing right.

This is similar to #87 by @lambdalemon but not as heavily refactored (yet)
and also includes tests.

* fix: lint and types

* fix: WTF-8, WindowS!

* chore: remove test files from coverage

* fix(docs): clarify YET AGAIN this unscaled text space nonsense

* fix: clarify and correct charspace and wordspace

* fix: add default ascent and descent for simple3.pdf

* test: verify that scaling is applied properly to vertical fonts

* docs: clarify and deprecate glyph_offset  and adv